Added the validation files to a different folder
