+++ views/users/login.handlebars
@@ -39,4 +39,6 @@
   </form>
 
 </section>
-<script src="/public/js/validation/login_validation.js"></script>
+<script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc="
+  crossorigin="anonymous"></script>
+<script type="module" src="/public/js/users/login_validation.js"></script>
