+++ views/users/signup.handlebars
@@ -36,4 +36,6 @@
     <p>Already have an account? <a href="/users/login" id="login-here-link">Login here</a></p>
   </form>
 </section>
-<script src="/public/js/validation/signup_validation.js"></script>
+<script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc="
+  crossorigin="anonymous"></script>
+<script type="module" src="/public/js/users/signup_validation.js"></script>
